Created setup for stage and scenes (not final) - Kelly
So far created a stackpane but the pane only has the gardenslots for now, my buttons for Pea plants and cherrybomb are not there yet. Gridpane has to be fixed still. Uploading what i have so far to give an idea of what the class names and scenes are called.